added about component
